+++ linux-user/cris/target_mman.h
@@ -5,7 +5,7 @@
  * arch/cris/include/arch-v32/arch/processor.h
  * TASK_SIZE               0xb0000000
  */
-#define TASK_UNMAPPED_BASE TARGET_PAGE_ALIGN(0xb0000000 / 3)
+#define TASK_UNMAPPED_BASE ((size_t) TARGET_PAGE_ALIGN(0xb0000000 / 3))
 
 /* arch/cris/include/uapi/asm/elf.h */
 #define ELF_ET_DYN_BASE    (TASK_UNMAPPED_BASE * 2)
