+++ linux-user/mips/target_mman.h
@@ -19,7 +19,7 @@
  * TASK_UNMAPPED_BASE         PAGE_ALIGN(TASK_SIZE / 3)
  */
 #define TASK_UNMAPPED_BASE \
-    TARGET_PAGE_ALIGN((1ull << TARGET_VIRT_ADDR_SPACE_BITS) / 3)
+    ((size_t) TARGET_PAGE_ALIGN((1ull << TARGET_VIRT_ADDR_SPACE_BITS) / 3))
 
 /* arch/mips/include/asm/elf.h */
 #define ELF_ET_DYN_BASE       (TASK_UNMAPPED_BASE * 2)
